CS-1799: Added comments into PlatformMessage header file

+++ libs/Board/Platform/src/PlatformMessage.cpp
@@ -43,7 +43,7 @@ const QString& PlatformMessage::jsonErrorString() const
     return data_->jsonErrorString_;
 }
 
-uint PlatformMessage::jsonErrorOffset() const
+ulong PlatformMessage::jsonErrorOffset() const
 {
     return data_->jsonErrorOffset_;
 }
